template_processor: Allow to skip files or directories
Templated file names can now contain expressions to control if the
file or directory should be generated or skipped. If the expression
evaluates to a non-empty string, that string is used as file or
directory name. If the expression evaluates to an empty string the
generation of that file or directory is skipped.

This can be used to provide custom files in the source repo that are
not overridden by a template.

For example the directory:
`addon/depends/windows/{{ "mingw" if not config.nomingw }}` is not
generated if `'nomingw': True` is added to the config.